requirements.txt
discord==2.3.2
discord.py==2.3.2
huggingface-hub==0.20.3
python-dotenv==1.0.0
pyttsx3==2.90
regex==2023.12.25
requests==2.31.0
soundfile==0.12.1
urllib3==2.0.7
PyNaCl==1.5.0
waitress==2.1.2
Flask==3.0.2